refactor to ride context
