+++ locomotion-sdk/src/Components/ShowMyLocationButton/index.js
@@ -1,11 +1,13 @@
 import React from 'react';
 import styled from 'styled-components';
-import Button from '../../../Components/Button';
+import Button from '../Button';
+import icon from '../../assets/center_btn.png';
 
 const Sight = styled.Image`
   height: 50px;
   width: 50px;
 `;
+
 const DisplayButton = styled(Button)`
   background-color: unset;
 `;
@@ -13,7 +15,7 @@ const DisplayButton = styled(Button)`
 export default styled(({ displayButton, onPress }) => (
   displayButton ? (
     <DisplayButton onPress={onPress} data-test-id="MyLocationButton">
-      <Sight source={require('../../../assets/center_btn.png')} resizeMode="contain" />
+      <Sight source={icon} resizeMode="contain" />
     </DisplayButton>
   ) : null
 ))`
